extract LatencyRecorder as a top-level public class
